CRUD'S completados de: Categoria, Proveedor, Producto, Cliente

+++ app/Http/Controllers/ProductoController.php
@@ -2,8 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use App\Http\Controllers\ProductoController;
+use App\Models\Producto;
 
 class ProductoController extends Controller
 {
@@ -15,25 +16,55 @@ class ProductoController extends Controller
 
     public function store(Request $request)
     {
-        $producto = Producto::create($request->all());
-        //holi
-        return response()->json($producto, 201);
+        $producto = new Producto();
+        $producto->nombre = $request->input('nombre');
+        $producto->stock = $request->input('stock');
+        $producto->fechaVencimiento = $request->input('fechaVencimiento');
+        $producto->precio = $request->input('precio');
+        $producto->proveedores_id = $request->input('proveedores_id');
+        $producto->categorias_id = $request->input('categorias_id');
+
+        $producto->save();
+        return response()->json(['mensaje' => 'Producto registrado con éxito'], 201);
     }
 
     public function show(string $id)
     {
-        return $producto;
+        $producto = Producto::find($id);
+
+        if (!$producto) {
+            return response()->json(['mensaje' => 'Producto no encontrado'], 404);
+        }
+        return response()->json([$producto], 200);
     }
 
     public function update(Request $request, string $id)
     {
-        $producto->update($request->all());
-        return response()->json($producto, 200);
+        $producto = Producto::find($id);
+
+        if (!$producto) {
+            return response()->json(['mensaje' => 'Producto no encontrado'], 404);
+        }
+        $producto->nombre = $request->input('nombre');
+        $producto->stock = $request->input('stock');
+        $producto->fechaVencimiento = $request->input('fechaVencimiento');
+        $producto->precio = $request->input('precio');
+        $producto->proveedores_id = $request->input('proveedores_id');
+        $producto->categorias_id = $request->input('categorias_id');
+        $producto->save();
+
+        return response()->json(['mensaje' => 'Producto actualizado con éxito'], 200);
     }
 
     public function destroy(string $id)
     {
+        $producto = Producto::find($id);
+
+        if (!$producto) {
+            return response()->json(['mensaje' => 'Producto no encontrado'], 404);
+        }
+
         $producto->delete();
-        return response()->json(null, 204);
+        return response()->json(['mensaje' => 'Producto eliminado con éxito'], 200);
     }
 }
